requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
appdirs==1.4.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
av==12.1.0
cffi==1.16.0
click==8.1.7
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
Cython==3.0.10
decorator==5.1.1
docker-pycreds==0.4.0
exceptiongroup==1.2.1
executing==2.0.1
ffmpeg-python==0.2.0
filelock==3.14.0
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.64.1
h5py==3.10.0
imageio==2.34.1
install==1.3.5
ipympl==0.9.3
ipywidgets==8.1.3
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
lazy_loader==0.4
librosa==0.10.1
llvmlite==0.42.0
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
ninja==1.11.1.1
numba==0.59.1
opencv-python==4.10.0.82
packaging==24.0
pandas==2.0.3
parso==0.8.4
pesq==0.0.4
pexpect==4.9.0
platformdirs==4.2.2
pooch==1.8.2
prompt_toolkit==3.0.46
protobuf==3.20.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pybind11==2.12.0
pycparser==2.22
pyDeprecate==0.3.2
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.2
pyroomacoustics==0.7.3
pystoi==0.4.1
python-dateutil==2.9.0.post0
python_speech_features==0.6
pytorch-lightning==1.8.0
pytz==2024.1
PyYAML==6.0.1
scikit-learn==1.5.0
scipy==1.10.1
sdeint==0.3.0
seaborn==0.13.2
sentry-sdk==2.5.0
setproctitle==1.3.3
setuptools==69.5.1
six==1.16.0
smmap==5.0.1
soundfile==0.12.1
soxr==0.3.7
stack-data==0.6.3
sympy==1.12.1
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
torch-ema==0.3
tqdm==4.66.1
traitlets==5.14.3
trampoline==0.1.2
triton==2.3.1
tzdata==2024.1
wandb==0.16.2
wcwidth==0.2.13
Werkzeug==3.0.3
wheel==0.43.0
widgetsnbextension==4.0.11
yarl==1.9.4
